backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
xmlsec>=1.3,<2.0
defusedxml>=0.7,<1.0
reportlab>=4.0,<5.0
openpyxl>=3.1,<4.0
python-dateutil>=2.8,<3.0
pytz>=2024.1,<2025.0
gunicorn>=21.0,<22.0